\name{jrmmisc-package}
\alias{jrmmisc-package}
\alias{jrmmisc}
\docType{package}
\title{
jrmmisc
}
\description{
Miscellaneous functions in the spirit of DRY...
}
\details{
\tabular{ll}{
Package: \tab jrmmisc\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-05-16\cr
License: \tab GPL (>=2) | BSD\cr
}
To do \ldots
}
\author{
John Minter

Maintainer: John Minter <jrminter@gmail.com>
}
\references{
TO DO
}
\keyword{ package }
\examples{
x <- c(1.1, 1.2, 1.05, 1.07, 0.99, 0.995, 0.98)
v.se <- v.std.err(x)
v.ucl <- v.u.cl(x)
v.lcl <- v.l.cl(x)
}
