\name{misclassification.from.table}
\alias{
misclassification.from.table
}
\title{
Compute the misclassification index from a table
}
\description{
Computes the misclassification index from a table, x.
}
\usage{
misclassification.from.table(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The table to process.
}
 
}
\value{
Returns misclassification index.
}
\author{
J. R Minter
}

\examples{
# to do
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{attribute}
